mac2lisa updates
mac2.lisa.py no longer requires the original Lisa ProFile image to retrieve the tag bytes from. For MacWorks, it seems that tag bytes of 0x00 work fine. The 20th byte is still set to the checksum of the block. The tag bytes of first block are also set properly to a boot the Lisa.